ActionEditor WIP (to create class) OSCModel save/load and OSCModelManagerUI addExisting getCommandDefinition fix

+++ Source/BaseItemEditor.h
@@ -31,7 +31,7 @@ public:
 	ScopedPointer<StringParameterUI> nameUI;
 	ScopedPointer<ImageButton> removeBT;
 
-	float headerHeight;
+	int headerHeight;
 
 	bool paintHeaderOnly;
